ChartAppBackend/chartapp-backend/requirements.txt
Django==5.1.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1